[zmarkdown/rebber-plugins]: reorganise files
